change data type

+++ model/contest_meta.go
@@ -13,7 +13,7 @@ type ContestMeta struct {
 	BeginTime                  time.Time
 	Duration                   uint16
 	NeededCorrectors           uint8
-	AllowedCorrectorUsageTimes uint8
+	AllowedCorrectorUsageTimes uint
 	AllowCorrectTilQuestion    uint8
 	NeededTickets              uint8
 }
